growth_studio/requirements/dev.txt
Django
Fabric3
pep8